+++ packages/gatsby-tinacms-json/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.0-alpha.2...gatsby-tinacms-json@0.8.0) (2019-12-02)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+# [0.8.0-alpha.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.0-alpha.1...gatsby-tinacms-json@0.8.0-alpha.2) (2019-12-02)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
 # [0.8.0-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.0-alpha.0...gatsby-tinacms-json@0.8.0-alpha.1) (2019-11-28)